move mirror to its own module
